ownpagen ja reservationsin komponenttien erittely
